+++ public/lib/url/url_canon_stdstring.h
@@ -42,7 +42,7 @@ class URL_EXPORT StdStringCanonOutput : public CanonOutput {
   // Must be called after writing has completed but before the string is used.
   void Complete();
 
-  void Resize(int sz) override;
+  void Resize(size_t sz) override;
 
  protected:
   std::string* str_;
