+++ public/lib/url/url_util.h
@@ -66,13 +66,13 @@ URL_EXPORT void LockStandardSchemes();
 // input (if any). The |compare| scheme must be a valid canonical scheme or
 // the result of the comparison is undefined.
 URL_EXPORT bool FindAndCompareScheme(const char* str,
-                                     int str_len,
+                                     size_t str_len,
                                      const char* compare,
                                      Component* found_scheme);
 inline bool FindAndCompareScheme(const std::string& str,
                                  const char* compare,
                                  Component* found_scheme) {
-  return FindAndCompareScheme(str.data(), static_cast<int>(str.size()),
+  return FindAndCompareScheme(str.data(), str.size(),
                               compare, found_scheme);
 }
 
@@ -93,7 +93,7 @@ URL_EXPORT bool IsStandard(const char* spec, const Component& scheme);
 // output and parsed structures will still be filled and will be consistent,
 // but they will not represent a loadable URL.
 URL_EXPORT bool Canonicalize(const char* spec,
-                             int spec_len,
+                             size_t spec_len,
                              bool trim_path_end,
                              CharsetConverter* charset_converter,
                              CanonOutput* output,
@@ -110,10 +110,10 @@ URL_EXPORT bool Canonicalize(const char* spec,
 // Returns true if the output is valid, false if the input could not produce
 // a valid URL.
 URL_EXPORT bool ResolveRelative(const char* base_spec,
-                                int base_spec_len,
+                                size_t base_spec_len,
                                 const Parsed& base_parsed,
                                 const char* relative,
-                                int relative_length,
+                                size_t relative_length,
                                 CharsetConverter* charset_converter,
                                 CanonOutput* output,
                                 Parsed* output_parsed);
@@ -123,7 +123,7 @@ URL_EXPORT bool ResolveRelative(const char* base_spec,
 // Escapes the given string as defined by the JS method encodeURIComponent. See
 // https://developer.mozilla.org/en/JavaScript/Reference/Global_Objects/encodeURIComponent
 URL_EXPORT void EncodeURIComponent(const char* input,
-                                   int length,
+                                   size_t length,
                                    CanonOutput* output);
 
 }  // namespace url
